+++ src/app/component-library/data-address/data-address-type-select/data-address-type-select-items.ts
@@ -6,6 +6,13 @@ export const dataAddressTypeSelectItems = (
 ): DataAddressTypeSelectItem[] => {
   const items: DataAddressTypeSelectItem[] = [];
 
+  if (type.startsWith('Datasource')) {
+    items.push({
+      id: 'On-Request',
+      label: '"On Request" Data Offer',
+    });
+  }
+
   items.push(
     {
       id: 'Http',