+++ src/app/routes/connector-ui/asset-page/asset-edit-dialog/form/model/temporal-coverage-form-model.ts
@@ -1,0 +1,15 @@
+import {FormControl, ɵFormGroupValue} from '@angular/forms';
+
+/**
+ * Form Model for AssetEditorDialog > Advanced > Temporal Coverage
+ */
+export interface TemporalCoverageFormModel {
+  from: FormControl<Date | null>;
+  toInclusive: FormControl<Date | null>;
+}
+
+/**
+ * Form Value for AssetEditorDialog > Advanced > Temporal Coverage
+ */
+export type TemporalCoverageFormValue =
+  ɵFormGroupValue<TemporalCoverageFormModel>;