+++ src/app/routes/connector-ui/asset-page/data-category-select/data-category-select-item.ts
@@ -1,0 +1,4 @@
+export interface DataCategorySelectItem {
+  id: string;
+  label: string;
+}