+++ src/app/routes/connector-ui/asset-page/language-select/language-select-item.ts
@@ -1,0 +1,7 @@
+export interface LanguageSelectItem {
+  id: string;
+  label: string;
+  comment?: string;
+  idShort?: string;
+  sameAs?: string;
+}